Unreviewed speculative build fix for clang.
Added explicit static_cast from int64_t to int32_t.

* platform/text/TextBreakIteratorICU.cpp:
(WebCore::uTextLatin1Clone):
(WebCore::uTextLatin1Extract):
(WebCore::uTextLatin1MapNativeIndexToUTF16):


Canonical link: https://commits.webkit.org/115710@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129698 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/text/TextBreakIteratorICU.cpp
@@ -102,7 +102,7 @@ static UText* uTextLatin1Clone(UText* destination, const UText* source, UBool de
     /* Point at the same position, but with an empty buffer */
     result->chunkNativeStart = source->chunkNativeStart;
     result->chunkNativeLimit = source->chunkNativeStart;
-    result->nativeIndexingLimit = source->chunkNativeStart;
+    result->nativeIndexingLimit = static_cast<int32_t>(source->chunkNativeStart);
     result->chunkOffset = 0;
     result->context = source->context;
     result->a = source->a;
@@ -200,7 +200,7 @@ static int32_t uTextLatin1Extract(UText* uText, int64_t start, int64_t limit, UC
         return 0;
 
     if (destCapacity > 0 && !dest) {
-        int32_t trimmedLength = length;
+        int32_t trimmedLength = static_cast<int32_t>(length);
         if (trimmedLength > destCapacity)
             trimmedLength = destCapacity;
 
@@ -216,7 +216,7 @@ static int32_t uTextLatin1Extract(UText* uText, int64_t start, int64_t limit, UC
     else
         *status = U_BUFFER_OVERFLOW_ERROR;
 
-    return length;
+    return static_cast<int32_t>(length);
 }
 
 static int64_t uTextLatin1MapOffsetToNative(const UText* uText)
@@ -228,7 +228,7 @@ static int32_t uTextLatin1MapNativeIndexToUTF16(const UText* uText, int64_t nati
 {
     ASSERT_UNUSED(uText, uText->chunkNativeStart >= nativeIndex);
     ASSERT_UNUSED(uText, nativeIndex < uText->chunkNativeLimit);
-    return nativeIndex;
+    return static_cast<int32_t>(nativeIndex);
 }
 
 static void uTextLatin1Close(UText* uText)
